move language-support/contextual/ast-builder to language-support/ast-viewer/ast-builder
